Preprocess scripts - calculate 450K windows
preprocess DNAme bulk and single cell data: define windows around CpGs measured by 450K array DNAme. For each bulk and\or sc sample the methylation level of a window is the mean methylation of all CpGs in this window. This code is used as preprocessing for the DNAme bulk-based classifier and the single cell DNAme clustering